refactor app by creating Routes component
